Aesthetic Design and Development of Humanoid Legged Robot
This paper presents a new full sized humanoid
leg robot that combines aesthetics and design theory with
practical research goals in robotics. The research goal of the
robot is to create human-like and compliant motion in multiple
contact situations through the use of torque controlled joints.
Low gear ratio and direct connections are used at each joint
for low-friction and back-drivability but without explicit joint
torque sensors. On the other hand, in creating human-like
motions, not only technical specifications but also aesthetic
design is important as the same performance of the robot
can be perceived very differently depending on the design.
The aesthetic design is, in this paper, achieved by the robot
design process using an integrated design and frame through
multi-axis CNC machining. The unique integration of the frame
and design also drastically reduces parts and complexity of
assembly for easy maintenance. In this paper, the design process
and features are presented with range of motion, weight,
and key aesthetic decisions. Compliant motion capability is
demonstrated by experimental results